refactor: Move time_config.py to units module
