+++ pyaerocom/io/__init__.py
@@ -34,5 +34,5 @@ from pyaerocom.io.cams2_83.reader import ReadCAMS2_83
 from .pyaro.read_pyaro import ReadPyaro
 from .pyaro.pyaro_config import PyaroConfig
 
-from . import helpers_units
+# from . import helpers_units
 from . import netcdf_fix
